chore(commands): move command to separate modules
