+++ src/commands/config/remove.js
@@ -1,0 +1,10 @@
+import { deleteProfileConfig } from "../../utils/config.js";
+
+export const removeCommandDefinition = [
+  "remove",
+  "Remove configuration",
+  {},
+  (args) => {
+    deleteProfileConfig(args.profile);
+  },
+];
