+++ src/commands/models/info.js
@@ -1,0 +1,36 @@
+import { prettyPrint } from "../../utils/print.js";
+
+export const infoCommandDefinition = [
+  "info <model>",
+  "Show detailed information about a model",
+  {
+    model: {
+      type: "string",
+    },
+  },
+  async (args) => {
+    const {
+      id,
+      name,
+      size,
+      description,
+      token_limit,
+      source_model,
+      family,
+      tasks,
+      tags,
+    } = await args.client.model({ id: args.model });
+
+    prettyPrint({
+      id,
+      name,
+      size,
+      description,
+      source_model: source_model?.id,
+      family,
+      token_limit,
+      tasks,
+      tags,
+    });
+  },
+];
