+++ src/commands/tunes/methods.js
@@ -1,0 +1,11 @@
+export const methodsCommandDefinition = [
+  "methods",
+  "List all tune methods",
+  {},
+  async (args) => {
+    const methods = await args.client.tuneMethods();
+    methods.forEach((method) => {
+      console.log(`${method.id} (${method.name})`);
+    });
+  },
+];
